+++ .env
@@ -2,6 +2,7 @@
 
 # General
 NEXT_PUBLIC_LOCALE='en-US'
+NEXT_PUBLIC_CANONICAL_URL='https://penny-ecommerce.vercel.app'
 
 # TakeShape
 NEXT_PUBLIC_TAKESHAPE_API_URL='https://api.takeshape.io/project/06ccc3dc-a9da-4f5b-9142-5a104db52ee3/production/graphql'
