+++ .env-example
@@ -3,6 +3,7 @@
 
 # General
 NEXT_PUBLIC_LOCALE='en-US'
+NEXT_PUBLIC_CANONICAL_URL=''
 
 # TakeShape
